Move NGMP update to the new uncoupled tick

+++ Code/Libraries/Source/WWVegas/WWLib/bittype.h
@@ -43,7 +43,7 @@
 
 typedef unsigned char	uint8;
 typedef unsigned short	uint16;
-typedef unsigned long	uint32;
+typedef unsigned int	uint32;
 typedef unsigned int    uint;
 
 typedef signed char		sint8;
